==> CII/2_DataPreparation_InfluenceArea.ipynb <==
import pandas as pd
import geopandas as gpd

# Create Influence Area for de polygon

df_predio = gpd.read_file("../DatosProcesados/Predio/Predio.shp")

# create a 400m buffer
df_predio_buffer_400 = df_predio.copy() 
df_predio_buffer_400 = df_predio_buffer_400.to_crs("EPSG:5347") 
df_predio_buffer_400.geometry = df_predio_buffer_400.buffer(400)
df_predio_buffer_400 = df_predio_buffer_400.to_crs("EPSG:4326")
df_predio_buffer_400.to_file("../DatosProcesados/AreaInfluencia/predio_buffer_400m.shp")

# create a 1200m buffer
df_predio_buffer_1200 = df_predio.copy() 
df_predio_buffer_1200 = df_predio_buffer_1200.to_crs("EPSG:5347") 
df_predio_buffer_1200.geometry = df_predio_buffer_1200.buffer(1200)
df_predio_buffer_1200 = df_predio_buffer_1200.to_crs("EPSG:4326")
df_predio_buffer_1200.to_file("../DatosProcesados/AreaInfluencia/predio_buffer_1200m.shp")

# Create Influence Area for travel time

# import network analysis result: 10min travel time
df_ttime_10min = gpd.read_file("../DatosProcesados/AreaInfluencia/traveltime_bus_10min.shp")

# concat with 1200m radius buffer
df_ia_1 = pd.concat([df_predio_buffer_1200[['geometry']], df_ttime_10min[['geometry']]])
# dissolver both geometries
df_ia_1['dis'] = 'x'
df_ia_1 = df_ia_1.dissolve(by='dis').reset_index()
# calculate area
df_ia_1 = df_ia_1.to_crs("EPSG:5347") 
df_ia_1['sup_m2'] = df_ia_1.area
df_ia_1 = df_ia_1.to_crs("EPSG:4326")

# save result
df_ia_1.to_file("../DatosProcesados/AreaInfluencia/predio_ttime10min_1200m.shp")
