==> CII/3_DataPreparation_NetworkStreets.ipynb <==
# CII proyect: Street Data Preparation For Network Analysis
### Process:
- Create street layer    
***Run topology of streets in ArcGIS***
- Create bus line buffer
- calculate travel time   
***Run Network Analyst in ArcGIS***

import pandas as pd
import geopandas as gpd

# import local goverments
df_muni = gpd.read_file("D:/JOB/UNSAM/2022 - CII/Datos/municipio/municipio.shp")
df_muni = df_muni.loc[df_muni.nam == 'La Matanza']

# Create street layer

# import streets
df_nac = gpd.read_file("../Datos/Calles/RedVialNacional.geojson")
df_pro = gpd.read_file("../Datos/Calles/RedVialProvincial.geojson")
df_urb = gpd.read_file("../Datos/Calles/RedVialUrbana.geojson")

# clean street data
df_urb['hct'] = df_urb['objeto']
df_calles = pd.concat([df_nac[['hct', 'geometry']], df_pro[['hct', 'geometry']], df_urb[['hct', 'geometry']]])
df_calles.loc[df_calles.hct == 2, 'hct'] = 'Provincial'
print(df_calles.hct.unique())
print(len(df_calles))
df_calles.head()

df_calles_lm = df_calles.sjoin(df_muni[['geometry']])
df_calles_lm.drop(columns='index_right', inplace=True)
print(df_calles_lm.hct.unique())
print(len(df_calles_lm))
df_calles_lm.head()

# save layer
df_calles_lm.to_file("../DatosProcesados/CallesLaMatanza.shp")

### Run topology of streets in ArcGIS

# Create bus lines buffer (10m)

# import bus lines
df_redcol = gpd.read_file("../DatosProcesados/Urbanos/RecorridoColectivos.geojson")
# create a bus line buffer
df_redcol_buffer = df_redcol.copy() 
df_redcol_buffer = df_redcol_buffer.to_crs("EPSG:3035") 
df_redcol_buffer.geometry = df_redcol_buffer.buffer(20)
df_redcol_buffer = df_redcol_buffer.to_crs("EPSG:4326")
df_redcol_buffer.to_file("../DatosProcesados/CallesNetwork/RecorridoColectivos.shp")

# Calculate travel time
- walking time: 80m/min
- bus time: 40km/h == 670m/min

df_calles_lm = gpd.read_file("../DatosProcesados/CallesNetwork/Calles_topo.shp")
df_redcol_buffer = gpd.read_file("../DatosProcesados/CallesNetwork/RecorridoColectivos.shp")

# calculate length
df_calles_lm = df_calles_lm.to_crs("EPSG:3035")
df_calles_lm['len_m'] = df_calles_lm.length
df_calles_lm = df_calles_lm.to_crs("EPSG:4326")

# spatial join street with bus lines
df_redcol_buffer['bus_p'] = 1
df_calles_lm_bu = df_calles_lm.sjoin(df_redcol_buffer[['bus_p', 'geometry']], how='left', predicate='within')
df_calles_lm_bu.drop(columns='index_right', inplace=True)
df_calles_lm_bu.reset_index(inplace=True)
df_calles_lm_bu.rename(columns={df_calles_lm_bu.columns[0]:'ddup'}, inplace=True)
df_calles_lm_bu.drop_duplicates(subset='ddup', inplace=True)
df_calles_lm_bu.drop(columns='ddup', inplace=True)
df_calles_lm_bu.reset_index(inplace=True, drop=True)
df_calles_lm_bu.head()

# calculate travel time for each street
df_calles_lm_bu['ttime_walking'] = df_calles_lm_bu['len_m'] / 80
df_calles_lm_bu['ttime_wbus'] = df_calles_lm_bu['len_m'] / 80
df_calles_lm_bu.loc[df_calles_lm_bu.bus_p.notnull(), 'ttime_wbus'] = df_calles_lm_bu['len_m'] / 670
df_calles_lm_bu.head()

# save result
df_calles_lm_bu.to_file("../DatosProcesados/CallesNetwork/Calles_traveltime.shp")

### Run Network Analyst in ArcGIS
