==> CII/4_CalculateInfluenceAreaVaribles.ipynb <==
import pandas as pd
import geopandas as gpd

df_b400 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_buffer_400m.shp")
df_b1200 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_buffer_1200m.shp")
df_ttim10min_b1200 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_ttime10min_1200m_manual.shp")
df_CiudadEvita = gpd.read_file("../QGIS_projects/Datos/AreaGeográfica/CiudadEvita.shp")
df_IsidroCasanova = gpd.read_file("../QGIS_projects/Datos/AreaGeográfica/IsidroCasanova.shp")
df_muni = gpd.read_file("D:/JOB/UNSAM/2022 - CII/Datos/municipio/municipio.shp")
df_muni = df_muni.loc[df_muni.nam == 'La Matanza', ['geometry']]
lista_areas = [[df_b400, 'buffer_400'], [df_b1200, 'buffer_1200'], [df_ttim10min_b1200, 't10min_b1200'], 
               [df_CiudadEvita, 'Ciudad_Evita'], [df_IsidroCasanova, 'Isidro_Casanova'], [df_muni, 'La_Matanza']]

# load data: 
# dumps
df_basca = gpd.read_file("../DatosProcesados/Ambientales/BasuralesACieloAbierto.geojson", driver='GeoJSON')
# villas
df_renabap = gpd.read_file("../DatosProcesados/Urbanos/BarriosRENABAP.geojson", driver='GeoJSON')
# green space
df_epv = gpd.read_file("../DatosProcesados/Urbanos/EspaciosVerdes.geojson", driver='GeoJSON')
# health
df_hosp = gpd.read_file("../DatosProcesados/Urbanos/Hospitales.geojson", driver='GeoJSON')
df_caps = gpd.read_file("../DatosProcesados/Urbanos/CentrosDeAtencion_Salud.geojson", driver='GeoJSON')
# education
df_edu_pub_ini = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Publico_Inicial.geojson", driver='GeoJSON')
df_edu_pub_pri = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Publico_Primaria.geojson", driver='GeoJSON')
df_edu_pub_sec = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Publico_Secundaria.geojson", driver='GeoJSON')
df_edu_pub_ter = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Publico_Terciaria.geojson", driver='GeoJSON')
df_edu_pri_pri = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Privado_Primaria.geojson", driver='GeoJSON')
df_edu_pri_sec = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Privado_Secundaria.geojson", driver='GeoJSON')
df_edu_pri_ter = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Privado_Terciaria.geojson", driver='GeoJSON')
df_edu_pri_ini = gpd.read_file("../DatosProcesados/Urbanos/EstEducativo_Privado_Inicial.geojson", driver='GeoJSON')
# sport and recreation
df_sport = gpd.read_file("../DatosProcesados/Urbanos/CentrosDeportes.geojson", driver='GeoJSON')
df_sport_buffer = df_sport.to_crs("EPSG:5347")
df_sport_buffer.geometry = df_sport_buffer.buffer(1000)
df_sport_buffer = df_sport_buffer.to_crs("EPSG:4326")
df_sport_buffer.to_file("../DatosProcesados/Urbanos/CentrosDeportes_buffer1000m.geojson", driver='GeoJSON')
df_comcenter = gpd.read_file("../DatosProcesados/Urbanos/CentrosComunitarios.geojson", driver='GeoJSON')
df_comcenter_buffer = df_comcenter.to_crs("EPSG:5347")
df_comcenter_buffer.geometry = df_comcenter_buffer.buffer(1000)
df_comcenter_buffer = df_comcenter_buffer.to_crs("EPSG:4326")
df_comcenter_buffer.to_file("../DatosProcesados/Urbanos/CentrosComunitarios_buffer1000m.geojson", driver='GeoJSON')
# transportation
df_redvial = gpd.read_file("../DatosProcesados/Urbanos/RedVial.geojson", driver='GeoJSON')
df_redcol = gpd.read_file("../DatosProcesados/Urbanos/RecorridoColectivos.geojson", driver='GeoJSON')
df_estcol = gpd.read_file("../DatosProcesados/Urbanos/EstacionesColectivos.geojson", driver='GeoJSON')
df_redffcc = gpd.read_file("../DatosProcesados/Urbanos/RedFerrocarril.geojson", driver='GeoJSON')
df_estffcc = gpd.read_file("../DatosProcesados/Urbanos/EstacionesFerrocarril.geojson", driver='GeoJSON')
# shops
df_shop = gpd.read_file("../DatosProcesados/Urbanos/LocalesComercios.geojson", driver='GeoJSON')
df_food = gpd.read_file("../DatosProcesados/Urbanos/LocalesComida.geojson", driver='GeoJSON')
df_econ = gpd.read_file("../DatosProcesados/Urbanos/BancosCajeros.geojson", driver='GeoJSON')
# land use
df_usosuelo = gpd.read_file("../DatosProcesados/Urbanos/UsosDeSuelo.geojson", driver='GeoJSON')
# secutity
df_secur = gpd.read_file("../DatosProcesados/Urbanos/EstablecimientosSeguridad.geojson", driver='GeoJSON')
# census data
df_census = gpd.read_file("../DatosProcesados/Sociodemograficos/DatosCensales.geojson", driver='GeoJSON')
df_census.rename(columns={'hogares':'hogar_num'}, inplace=True)

df_muni = df_muni.to_crs("EPSG:5347")
df_muni_area = df_muni.area
df_muni = df_muni.to_crs("EPSG:4326")
df_muni_area

df_resumen = ''
for area in lista_areas:    
    df_area = area[0].copy()
    df_area['nombre'] = area[1]
    # count dumps
    df_area['basural_ca'] = len(df_area.sjoin(df_basca))
    # data villas
    df_area['barrios_pop'] = len(df_area.sjoin(df_renabap))
    df_area['barrios_pop_flias'] = df_area.sjoin(df_renabap)['cantidad_familias_aproximada'].sum()
    df_area['barrios_pop_viv'] = df_area.sjoin(df_renabap)['cantidad_viviendas_aproximadas'].sum()
    df_energia = df_area.sjoin(df_renabap)[['energia_electrica']]
    df_area['barrios_pop_porc_conex_energia'] = round(len(df_energia.loc[df_energia.energia_electrica.str.contains('Conex')])/len(df_energia)*100, 2)
    df_agua = df_area.sjoin(df_renabap)[['agua_corriente']]
    df_area['barrios_pop_porc_conex_agua'] = round(len(df_agua.loc[df_agua.agua_corriente.str.contains('Conex')])/len(df_agua) *100, 2)
    # green spaces
    df_area['evp'] = len(df_area.sjoin(df_epv))
    df_area['evp_sup_m2'] = df_area.sjoin(df_epv)['sup_m2'].sum()
    df_area['evp_sup_porc'] = round(100 * df_area['evp_sup_m2'] / df_epv.sup_m2.sum(), 2)
    df_area['evp_supLM_porc'] = round(100 * df_area['evp_sup_m2'] / list(df_muni_area)[0], 2)
    # health
    df_area['salud_hosp_pub'] = len(df_area.sjoin(df_hosp))
    df_area['salud_caps_pub'] = len(df_area.sjoin(df_caps))
    # education
    df_area['edu_inicial_pub'] = len(df_area.sjoin(df_edu_pub_ini))
    df_area['edu_primaria_pub'] = len(df_area.sjoin(df_edu_pub_pri))
    df_area['edu_secundaria_pub'] = len(df_area.sjoin(df_edu_pub_sec))
    df_area['edu_terciaria_pub'] = len(df_area.sjoin(df_edu_pub_ter))
    df_area['edu_inicial_pri'] = len(df_area.sjoin(df_edu_pri_ini))
    df_area['edu_primaria_pri'] = len(df_area.sjoin(df_edu_pri_pri))
    df_area['edu_secundaria_pri'] = len(df_area.sjoin(df_edu_pri_sec))
    df_area['edu_terciaria_pri'] = len(df_area.sjoin(df_edu_pri_ter))

    df_area['edu_inicial'] = df_area['edu_inicial_pub'] + df_area['edu_inicial_pri']
    df_area['edu_primaria'] = df_area['edu_primaria_pub'] + df_area['edu_primaria_pri']
    df_area['edu_secundaria'] = df_area['edu_secundaria_pub'] + df_area['edu_secundaria_pri']
    df_area['edu_terciaria'] = df_area['edu_terciaria_pub'] + df_area['edu_terciaria_pri']
    # sport and recreation
    df_area['esp_deportivos'] = len(df_area.sjoin(df_sport))
    df_area_sport_buffer = df_area.to_crs("EPSG:5347")
    sup_tot_m2_sport_buffer = df_area_sport_buffer.area.sum()
    df_area_sport_buffer = df_area_sport_buffer.to_crs("EPSG:4326")
    df_area_sport_buffer = df_area_sport_buffer.overlay(df_sport_buffer, how='difference')
    df_area_sport_buffer = df_area_sport_buffer.to_crs("EPSG:5347")
    sup_dif_m2_sport_buffer = df_area_sport_buffer.area.sum()
    df_area['esp_deportivos_porc1000m'] = round(100- 100* sup_dif_m2_sport_buffer / sup_tot_m2_sport_buffer, 2)

    df_area['centros_comuni'] = len(df_area.sjoin(df_comcenter))
    df_area_comuni_buffer = df_area.to_crs("EPSG:5347")
    sup_tot_m2_comuni_buffer = df_area_comuni_buffer.area.sum()
    df_area_comuni_buffer = df_area_comuni_buffer.to_crs("EPSG:4326")
    df_area_comuni_buffer = df_area_comuni_buffer.overlay(df_comcenter_buffer, how='difference')
    df_area_comuni_buffer = df_area_comuni_buffer.to_crs("EPSG:5347")
    sup_dif_m2_comuni_buffer = df_area_comuni_buffer.area.sum()
    df_area['centros_comuni_porc1000m'] = round(100- 100* sup_dif_m2_comuni_buffer / sup_tot_m2_comuni_buffer, 2)

    # education
    df_area['red_vial'] = len(df_area.sjoin(df_redvial)['fna1'].unique())
    df_area['colec_red'] = len(df_area.sjoin(df_redcol)['NUM_LI'].unique())
    df_area['colec_paradas'] = len(df_area.sjoin(df_estcol))
    df_area['ffcc_red'] = len(df_area.sjoin(df_redffcc)['nam'].unique())
    df_area['ffcc_estac'] = len(df_area.sjoin(df_estffcc))
    # shops
    df_area['comer_local'] = len(df_area.sjoin(df_shop))
    df_area['comer_comida'] = len(df_area.sjoin(df_food))
    df_area['comer_econo'] = len(df_area.sjoin(df_econ))
    # land use
    df_usosuelo_calc = gpd.overlay(df_usosuelo, df_area, how='intersection').groupby('descripcio').sum('sup_m2').reset_index()[['descripcio', 'sup_m2']]
    df_usosuelo_calc['sup_km2'] = round(df_usosuelo_calc['sup_m2'] / 1000000, 2)
    df_usosuelo_calc['sup_porc'] = round(100*df_usosuelo_calc['sup_m2'] / df_usosuelo_calc.sup_m2.sum(), 2)
    df_usosuelo_calc.sort_values(by='sup_porc', ascending=False, inplace=True)
    df_usosuelo_calc.reset_index(drop=True, inplace=True)
    str_usosuelo = ""
    for i in range(len(df_usosuelo_calc)):
        if str_usosuelo != "":
            str_usosuelo += " -- "
        str_usosuelo += df_usosuelo_calc.loc[i, 'descripcio'] +": "+  str(df_usosuelo_calc.loc[i, 'sup_km2']) +"km² ("+ str(df_usosuelo_calc.loc[i, 'sup_porc']) +"%)"
    df_area['uso_suelo'] = str_usosuelo
    # security
    df_area['est_seguridad'] = len(df_area.sjoin(df_secur))
    # census
    df_area['habitantes'] = df_area.sjoin(df_census)['personas'].sum()
    df_area['habitantes_varon'] = df_area.sjoin(df_census)['varon'].sum()
    df_area['habitantes_mujer'] = df_area.sjoin(df_census)['mujer'].sum()
    df_area['hogares'] = df_area.sjoin(df_census)['hogar_num'].sum()
    df_area['densidad'] = round(df_area.sjoin(df_census)['personas'].sum() / (df_area.sjoin(df_census)['sup_m2'].sum()/1000000),0)
    df_area['nbi_habitantes'] = df_area.sjoin(df_census)['nbi_pers'].sum()
    df_area['nbi_habitantes_porc'] = round(100* df_area.sjoin(df_census)['nbi_pers'].sum() / df_area.sjoin(df_census)['personas'].sum(), 2)
    df_area['nbi_hogares'] = df_area.sjoin(df_census)['nbi_hog'].sum()
    df_area['nbi_hogares_porc'] = round(100* df_area.sjoin(df_census)['nbi_hog'].sum() / df_area.sjoin(df_census)['hogar_num'].sum(), 2)

    if len(df_resumen) == 0:
        df_resumen = df_area
    else:
        df_resumen = pd.concat([df_resumen, df_area])

df_resumen['barrios_pop_porc'] = round(100* df_resumen['barrios_pop'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'barrios_pop'])[0], 2)
df_resumen['salud_hosp_pub_porc'] = round(100* df_resumen['salud_hosp_pub'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'salud_hosp_pub'])[0], 2)
df_resumen['salud_caps_pub_porc'] = round(100* df_resumen['salud_caps_pub'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'salud_caps_pub'])[0], 2)
df_resumen['edu_inicial_porc'] = round(100* df_resumen['edu_inicial'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'edu_inicial'])[0], 2)
df_resumen['edu_primaria_porc'] = round(100* df_resumen['edu_primaria'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'edu_primaria'])[0], 2)
df_resumen['edu_secundaria_porc'] = round(100* df_resumen['edu_secundaria'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'edu_secundaria'])[0], 2)
df_resumen['edu_terciaria_porc'] = round(100* df_resumen['edu_terciaria'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'edu_terciaria'])[0], 2)
df_resumen['esp_deportivos_porc'] = round(100* df_resumen['esp_deportivos'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'esp_deportivos'])[0], 2)
df_resumen['centros_comuni_porc'] = round(100* df_resumen['centros_comuni'] / list(df_resumen.loc[df_resumen.nombre == 'La_Matanza', 'centros_comuni'])[0], 2)
df_resumen['evp_sup_m2_hab'] = round(100* df_resumen['evp_sup_m2'] / df_resumen['habitantes'], 2)

df_resumen_trasp = df_resumen.T.reset_index()
df_resumen_trasp.columns = list(df_resumen_trasp.loc[2])
df_resumen_trasp = df_resumen_trasp.loc[3:].reset_index(drop=True)
df_resumen_trasp['descripcion'] = 'x'
df_resumen_trasp = df_resumen_trasp[['nombre', 'descripcion', 'buffer_400', 'buffer_1200', 't10min_b1200', 'Ciudad_Evita', 'Isidro_Casanova', 'La_Matanza']]
df_resumen_trasp.head()

dict_columns = {
    'basural_ca':'cantidad de basurales a cielo abierto', 
    'barrios_pop':'cantidad de barrios populares',
    'barrios_pop_porc':'porcentaje de barrios populares respecto a La Matanza',
    'barrios_pop_flias':'cantidad de familias en barrios populares', 
    'barrios_pop_viv':'cantidad de viviendas en barrios populares',
    'barrios_pop_porc_conex_energia':'cantidad de barrios populares con conexión a energía elecrtica', 
    'barrios_pop_porc_conex_agua':'cantidad de barrios populares con conexión a red de agua', 
    'evp':'cantidad de espacios verdes públicos',
    'evp_sup_m2':'superficie total de espacios verdes en m²', 
    'evp_sup_porc':'porcentaje de superficie de espacios verdes respecto a espacios verdes de La Matanza', 
    'evp_supLM_porc':'porcentaje de superficie de espacios verdes respecto a la superficie del partido de La Matanza', 
    'evp_sup_m2_hab':'superficie total de espacios verdes en m² por habitante',
    'salud_hosp_pub':'cantidad de hospitales públicos', 
    'salud_hosp_pub_porc':'porcentaje de hospitales públicos respecto a La Matanza', 
    'salud_caps_pub':'cantidad de centros de atención primaria',
    'salud_caps_pub_porc':'porcentaje de centros de atención primaria respecto a La Matanza',
    'edu_inicial':'cantidad de establecimientos de nivel inicial', 
    'edu_primaria':'cantidad de establecimientos de nivel primario', 
    'edu_secundaria':'cantidad de establecimientos de nivel secundario',
    'edu_terciaria':'cantidad de establecimientos de nivel terciario', 
    'edu_inicial_porc':'porcentaje de establecimientos de nivel inicial respecto a La Matanza', 
    'edu_primaria_porc':'porcentaje de establecimientos de nivel primario respecto a La Matanza', 
    'edu_secundaria_porc':'porcentaje de establecimientos de nivel secundario respecto a La Matanza',
    'edu_terciaria_porc':'porcentaje de establecimientos de nivel terciario respecto a La Matanza',
    'edu_inicial_pub':'cantidad de establecimientos de nivel inicial públicos', 
    'edu_primaria_pub':'cantidad de establecimientos de nivel primario públicos', 
    'edu_secundaria_pub':'cantidad de establecimientos de nivel secundario públicos',
    'edu_terciaria_pub':'cantidad de establecimientos de nivel terciario públicos', 
    'edu_inicial_pri':'cantidad de establecimientos de nivel inicial privado', 
    'edu_primaria_pri':'cantidad de establecimientos de nivel primario privado',
    'edu_secundaria_pri':'cantidad de establecimientos de nivel secundario privado', 
    'edu_terciaria_pri':'cantidad de establecimientos de nivel terciario privado', 
    'esp_deportivos':'cantidad de establecimientos deportivos',
    'esp_deportivos_porc':'porcentaje de establecimientos deportivos respecto a La Matanza',
    'esp_deportivos_porc1000m':'porcentaje de superficie no cubierta a 1000m de establecimientos deportivos',
    'centros_comuni':'cantidad de centros comunitarios',
    'centros_comuni_porc':'porcentaje de centros comunitarios respecto a La Matanza', 
    'centros_comuni_porc1000m':'porcentaje de superficie no cubierta a 1000m de centros comunitarios',
    'red_vial':'cantidad de redes viales de acceso', 
    'colec_red':'cantida de líneas de colectivo que cruzan el área', 
    'colec_paradas':'cantidad de paradas de colectivos', 
    'ffcc_red':'cantidad de líneas de ferrocarriles que cruzan el área',
    'ffcc_estac':'cantidad de estaciones de ferrocarriles', 
    'comer_local':'cantidad de establecimientos comerciales', 
    'comer_comida':'cantidad de establecimientos de comida (restaurant, fast-food, etc)', 
    'comer_econo':'cantidad de establecimientos de actividad económica (bancos, cajeros)', 
    'uso_suelo':'uso de suelo: tipo de uso - superficie en km² - porcentaje sobre el total',
    'est_seguridad':'cantidad de establecimientos de seguridad', 
    'habitantes':'cantidad de habitantes según censo', 
    'habitantes_varon':'cantidad de habitantes varones', 
    'habitantes_mujer':'cantidad de habitantes mujeres',
    'hogares':'cantidad de hogares', 
    'densidad':'densidad poblacional', 
    'nbi_habitantes':'cantidad de habitantes con nbi', 
    'nbi_habitantes_porc':'porcentaje de habitantes con nbi',
    'nbi_hogares':'cantidad de hogares con nbi',
    'nbi_hogares_porc':'porcentaje de hogares con nbi'
}

for i in dict_columns:
    df_resumen_trasp.loc[df_resumen_trasp.nombre == i, 'descripcion'] = dict_columns[i]

df_resumen_trasp.head()

df_resumen_trasp.to_csv("../DatosProcesados/Tablas/Datos_AreasInfluencia.csv", encoding='utf8', sep=',', decimal='.')
