==> CII/5_Datos_CABA.ipynb <==
import pandas as pd
import geopandas as gpd

df_CABA = gpd.read_file("../QGIS_projects/Datos/CiudadBA.shp")
df_AMBA = gpd.read_file("../QGIS_projects/Datos/AMBA.shp")
df_censo2010 = gpd.read_file("../Datos/CABA/informacion_censal_por_radio_2010_wgs84.shp")
df_evp = gpd.read_file("../Datos/CABA/espacio_verde_publico_wgs84.shp")
df_pobla = gpd.read_file("D:/JOB/UNSAM/2022 - CII/Datos/Sociodemografico/PBA_poblacion_y_datos_basicos_2010.shp")

# Datos CABA

# calculate surface
df_censo2010 = df_censo2010.to_crs("EPSG:5347")
df_censo2010['sup_k2'] = df_censo2010.area / 1000000
df_censo2010 = df_censo2010.to_crs("EPSG:4326")

hobitantes_tot = df_censo2010.TOTAL_POB.sum()
sup_tot_k2 = df_censo2010.sup_k2.sum()
densidad = int(round(hobitantes_tot / sup_tot_k2, 0))
hogares_tot = int(df_censo2010.T_HOGAR.sum())
hogares_nbi = int(df_censo2010.H_CON_NBI.sum())
hogares_nbi_porc = round(100* hogares_nbi / hogares_tot, 2)

evp_sup_m2 = df_evp['area'].sum()
evp_sup_m2_hab = round(evp_sup_m2 / hobitantes_tot, 2)

print("DATOS CABA")
print("Habitantes totales:", hobitantes_tot)
print("Densidad [hab/km²]:", densidad)
print("Hogares totales:", hogares_tot)
print("Hogares con NBI:", hogares_nbi)
print("Porcentaje de hogares con NBI:", hogares_nbi_porc)
print("Superficie Espacios Verdes [m²]:", evp_sup_m2)
print("Superficie Espacios Verdes por habitantes [m²/hab]:", evp_sup_m2_hab)

# Datos AMBA

df_AMBA_censo = df_AMBA.sjoin(df_pobla[['personas', 'geometry']])

AMBA_habitantes_tot = int(df_AMBA_censo.personas.sum())
print("DATOS AMBA")
print("Habitantes totales:", AMBA_habitantes_tot)
