==> CII/6_Capas_de_Datos.ipynb <==
import pandas as pd
import geopandas as gpd

df_b400 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_buffer_400m.shp")
df_b1200 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_buffer_1200m.shp")
df_ttim10min_b1200 = gpd.read_file("../DatosProcesados/AreaInfluencia/predio_ttime10min_1200m_manual.shp")
df_CiudadEvita = gpd.read_file("../QGIS_projects/Datos/CiudadEvita.shp")
df_IsidroCasanova = gpd.read_file("../QGIS_projects/Datos/IsidroCasanova.shp")
df_muni = gpd.read_file("D:/JOB/UNSAM/2022 - CII/Datos/municipio/municipio.shp")
df_muni = df_muni.loc[df_muni.nam == 'La Matanza', ['geometry']]
lista_areas = [[df_b400, 'buffer_400'], [df_b1200, 'buffer_1200'], [df_ttim10min_b1200, 't10min_b1200'], 
               [df_CiudadEvita, 'Ciudad_Evita'], [df_IsidroCasanova, 'Isidro_Casanova'], [df_muni, 'La_Matanza']]

# Densidad y NBI en Área de Influencia

# census data
df_census = gpd.read_file("../DatosProcesados/Sociodemograficos/DatosCensales.geojson", driver='GeoJSON')

df_ttim10min_b1200_intersection = df_ttim10min_b1200.overlay(df_census, how='intersection')
df_ttim10min_b1200_intersection['sup_k2'] = df_ttim10min_b1200_intersection['sup_m2'] / 1000000
df_ttim10min_b1200_intersection['densidad_habkm2'] = round(df_ttim10min_b1200_intersection['personas']/df_ttim10min_b1200_intersection['sup_k2'],0)
df_ttim10min_b1200_intersection.to_file("../QGIS_projects/Datos/Datos_inter/AI_datos_censales.geojson", driver='GeoJSON')

# Centros comunitarios y de deporte dentro de 1200m

# sport and recreation
df_sport = gpd.read_file("../DatosProcesados/Urbanos/CentrosDeportes.geojson", driver='GeoJSON')
df_comcenter = gpd.read_file("../DatosProcesados/Urbanos/CentrosComunitarios.geojson", driver='GeoJSON')

df_b1200_sport = df_sport.sjoin(df_b1200[['geometry']])
df_b1200_sport.drop(columns='index_right', inplace=True)
df_b1200_sport.to_file("../QGIS_projects/Datos/Datos_inter/AI_datos_centros_deporte.geojson", driver='GeoJSON')

""" df_b1200_comcenter = df_comcenter.sjoin(df_b1200[['geometry']])
df_b1200_comcenter.drop(columns='index_right', inplace=True)
df_b1200_comcenter.to_file("../QGIS_projects/Datos/Datos_inter/AI_datos_centros_comunitarios.geojson", driver='GeoJSON') """
